stdlib: Fix append_kernel_arg for RiscvBoard
Previously, the device tree was being generated when the workload was
initially set in `set_kernel_disk_workload` (when the disk image was
set). However, this meant that if you changed the kernel boot parameters
after the initial call they would not show up in the dtb file. The
RISC-V kernel picks up the parameters from the dtb file, not what is
passed in via the gem5 workload.

The solution is to generate the dtb file in `_pre_instantiate` instead
of in `_add_disk_to_board`

Signed-off-by: Jason Lowe-Power <jason@lowepower.com>