sim,util: External Hypercall signal 'SIGRTMIN' -> 'SIGHUP'
MacOS does not support real-time signals like 'SIGRTMIN'. Using this
makes "src/sim/init_signals.cc" uncompilable on Mac systems and would
make use of the external hypercall transmitter impossible for Mac users.

Given 'SIGUSR1' and 'SIGUSR2' are already used in gem5, this patch
changes the signal to 'SIGHUP'; a signal typically used to prompt
system daeomons to reload their configuration files. As this doesn't
apply to gem5, it seems save to use in this manner.